+++ src/arch/linux/x86_64/sigaction.h
@@ -14,14 +14,14 @@ static inline int32_t linux_x8664_sigaction(const int32_t signum, struct sigacti
     act->sa_flags |= SA_RESTORER;
     act->sa_restorer = &linux_x8664_restore_rt;
 
-    long ret = linux_x8664_asm_syscall4(
+    int32_t ret = linux_x8664_asm_syscall4(
         __NR_rt_sigaction,
         signum,
         act,
         oldact,
         sizeof(act->sa_mask));
 
-    if ((unsigned long)ret >= (unsigned long)-4095) {
+    if ((uint64_t)ret >= (uint64_t)-4095) {
         errno = -ret;
         return -1;
     }
